+++ src/Bar/Bar.vue
@@ -15,6 +15,7 @@
 import BarButtonGeneric from './BarButtonGeneric.vue'
 import BarButtonColor from './BarButtonColor.vue'
 import BarSeparator from './BarSeparator.vue'
+import BarSpacer from './BarSpacer.vue'
 
 import 'material-design-icons/iconfont/material-icons.css'
 
@@ -23,7 +24,8 @@ export default {
   components: {
     BarButtonGeneric,
     BarButtonColor,
-    BarSeparator
+    BarSeparator,
+    BarSpacer
   },
 
   props: {
@@ -44,9 +46,11 @@ export default {
       if(!this.$el.contains(e.target)) this.menu_open = false;
     },
     toggle_menu(name, event) {
+      event.stopPropagation();
       const ref = this.$refs[name][0];
+      const disabled = ref.item && ref.item.disabled;
       const touch = event.sourceCapabilities && event.sourceCapabilities.firesTouchEvents;
-      this.menu_open = ref.is_menu ? (touch ? true : !this.menu_open) : false;
+      this.menu_open = ref.is_menu && !disabled ? (touch ? true : !this.menu_open) : false;
     },
     get_component(is) {
       if(is && !Array.isArray(is) && typeof is == "object") return is;