+++ src/Bar/BarSpacer.vue
@@ -1,0 +1,9 @@
+<template>
+  <div class="bar-spacer"></div>
+</template>
+
+<style scoped>
+.bar-spacer {
+  flex-grow: 1;
+}
+</style>